query-service/requirements.txt
flask>=3.0
gunicorn>=22.0
google-cloud-bigquery>=3.0
